MOVE: Methods to extensions
