+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/Inputs.kt
@@ -13,6 +13,8 @@
 
 package slatekit.common
 
+import slatekit.common.ext.splitToMapOfType
+
 
 interface InputsUpdateable {
     // Immutable add
