+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/Strings.kt
@@ -21,18 +21,6 @@ package slatekit.common
 object Strings {
 
     /**
-     * returns a new string that truncates the string supplied by the count of chars.
-     */
-    fun truncate(text: String, count: Int): String {
-        return when {
-            text.isNullOrEmpty() -> text
-            text.length <= count -> text
-            else -> text.substring(0, count)
-        }
-    }
-
-
-    /**
      * Interprets a string which could have a macro function inside it.
      * NOTE: This is currently used for decryption of config settings at runtime
      * @param value   : "@{decrypt('abc123')"
@@ -58,19 +46,6 @@ object Strings {
     fun decrypt(value: String, decryptor: ((String) -> String)? = null): String =
             Strings.interpret(value, "decrypt", decryptor)
 
-
-
-
-    /**
-     * String match factoring in the wildcard "*"
-     */
-    fun isMatchOrWildCard(actual: String, expected: String, wildcard:String = "*"): Boolean {
-        return if (actual.isNullOrEmpty() || actual == wildcard)
-            true
-        else
-            actual == expected
-    }
-
 }
 
 
@@ -87,83 +62,6 @@ fun String.repeatWith(delimiter: String, count: Int): String {
  */
 fun String.pascalCase(): String {
     return this[0].toUpperCase() + this.substring(1)
-}
-
-/**
- * Splits the string to a map using the converters supplied
- * "a,b"     = map( "a" to "a", "b" to "b" )
- * "a=1,b=2" = map( "a" to 1  , "b" to 2 )
- */
-fun String.splitToMapOfType(
-    delimeterPairs: Char = ',',
-    trim: Boolean = true,
-    delimiterValue: Char? = '=',
-    keyConverter: ((String) -> Any)? = null,
-    valConverter: ((String) -> Any)? = null
-): Map<*, *> {
-    return if (this.isNullOrEmpty()) {
-        mapOf<Any, Any>()
-    } else {
-        val pairs = this.split(delimeterPairs)
-        val map = mutableMapOf<Any, Any>()
-        for (pair in pairs) {
-            val keyVal: Pair<String, String> = delimiterValue?.let { d ->
-                val tokens = pair.split(d)
-                Pair(tokens[0], tokens[1])
-            } ?: Pair(pair, pair)
-
-            val pkey = if (trim) keyVal.first.trim() else keyVal.first
-            val pval = if (trim) keyVal.second.trim() else keyVal.second
-            val finalKey = keyConverter?.let { k -> k(pkey) } ?: pkey
-            val finalVal = valConverter?.let { c -> c(pval) } ?: pval
-            map.put(finalKey, finalVal)
-        }
-        map.toMap()
-    }
-}
-
-/**
- * Splits the string to a map
- * "a,b"     = map( "a" to "a", "b" to "b" )
- * "a=1,b=2" = map( "a" to 1  , "b" to 2 )
- */
-fun String.splitToMapWithPairs(delimiterPairs: Char = ',', delimiterKeyValue: Char = '=', trim: Boolean = true): Map<String, String> {
-    val map = mutableMapOf<String, String>()
-    if (!this.isNullOrEmpty()) {
-        val pairs = this.split(delimiterPairs)
-        for (pair in pairs) {
-            val finalPair = if (trim) pair.trim() else pair
-            val tokens = finalPair.split(delimiterKeyValue)
-            val key = if (trim) tokens[0].trim() else tokens[0]
-
-            val kval = if (tokens.size > 1) {
-                if (trim) tokens[1].trim() else tokens[1]
-            } else {
-                key
-            }
-            map[key] = kval
-        }
-    }
-    return map.toMap()
-}
-
-/**
- * Gets a substring as a single Pair.
- * @example:
- * "user://appdir/log.txt".subString("://") -> Pair("user", "appdir/log.txt")
- */
-fun String.subStringPair(pattern: String): Pair<String, String>? {
-    return if (!this.isNullOrEmpty() && !pattern.isNullOrEmpty()) {
-        val ndxPattern = this.indexOf(pattern)
-        if (ndxPattern < 0) {
-            null
-        } else {
-            val part1 = this.substring(0, ndxPattern + pattern.length)
-            val remainder = this.substring(ndxPattern + pattern.length)
-            Pair(part1, remainder)
-        }
-    } else
-        null
 }
 
 val String.toCharMap: Map<Char, Boolean> get() = this.toCharArray().map { c -> c to true }.toMap()
