+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/Vars.kt
@@ -13,6 +13,7 @@
 
 package slatekit.common
 
+import slatekit.common.ext.splitToMapWithPairs
 import slatekit.common.utils.ListMap
 
 /**
