+++ src/lib/kotlin/slatekit-generator/src/main/kotlin/slatekit/generator/GeneratorContext.kt
@@ -1,6 +1,6 @@
 package slatekit.generator
 
-import slatekit.common.toId
+import slatekit.common.ext.toId
 import java.io.File
 
 /**
